refactor(tui): extract the reasoning region out of input.rs
`input.rs` is over the code-size budget, and the panic fixes pushed it further
over (4016 -> 4048 LOC), which CI's oversized-file ratchet correctly rejected:
existing oversized files must shrink or stay flat.

Rather than rebaseline the ratchet, move the streaming reasoning-region methods
into their own module. They belong together for a real reason, now documented at
the top of the new file: `reasoning_partial_len` and `reasoning_block_start` are
both byte offsets into `streaming.streaming_text` that are recorded at one point
and used to slice the buffer later, which is precisely the shape that produced
the #632 crash twice. Naming that invariant in one place is what makes the next
instance of this bug findable.

No behavior change. `input.rs` drops to 3857 LOC, 159 below its baseline, so it
now satisfies the ratchet rather than merely being excused from it.

refs #632
